Removed unused files, cleaned up unused types Renamed SimpleRevealManager to RevealManager Renamed RevealManager to RevealManagerBase
