--- viewer/src/data/provider/BinaryBlobFileProvider.ts
@@ -1,7 +1,0 @@
-/*!
- * Copyright 2020 Cognite AS
- */
-
-export interface BinaryBlobFileProvider {
-  fetchBinaryFileByPath(blobId: number, filePath: string): Promise<Uint8Array>;
-}
