--- viewer/src/data/provider/BinaryFileProvider.ts
@@ -1,8 +1,0 @@
-/*!
- * Copyright 2020 Cognite AS
- */
-
-export interface BinaryFileProvider {
-  fetchBinaryFileFromUrl(url: string): Promise<Uint8Array>;
-  fetchBinaryFileFromCdf(blobId: number, filePath: string): Promise<Uint8Array>;
-}
