+++ viewer/src/views/threejs/index.ts
@@ -10,9 +10,9 @@ export { RootSectorNode } from './cad/RootSectorNode';
 export { ModelNodeAppearance } from '../common/cad/ModelNodeAppearance';
 export { GlobalNodeAppearance } from '../common/cad/GlobalNodeAppearance';
 
-export { SimpleRevealManager } from '../../component/three/SimpleRevealManager';
-
 export { RevealManager } from '../../component/three/RevealManager';
+
+export { RevealManagerBase } from '../../component/three/RevealManagerBase';
 export { MaterialManager } from './cad/MaterialManager';
 export { CadSectorParser } from '../../data/parser/CadSectorParser';
 export { SimpleAndDetailedToSector3D } from '../../data/transformer/three/SimpleAndDetailedToSector3D';
